+++ CHANGES.rst
@@ -1,3 +1,8 @@
+Next release
+============
+
+* [ENH] Use robust template generation for multiple T1w images (#481)
+
 0.4.1 (20th of April 2017)
 ==========================